pyproject.toml
[project]
name = "docuweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
]
